Revert "merge master"
This reverts commit 544fa1806edbba8d975812e4f431793fc85a76ac, reversing
changes made to 12fc0491eca14e18f47931d031fb6ad2d7f18e33.

revertir fusion con master

+++ i386/i386/idt.c
@@ -30,7 +30,7 @@ struct real_gate idt[IDTSZ];
 
 struct idt_init_entry
 {
-	unsigned long entrypoint;
+	unsigned entrypoint;
 	unsigned short vector;
 	unsigned short type;
 };
